+++ MapTB-Article.h
@@ -9,7 +9,7 @@ namespace NS_Comp_Mappage9
 		System::String^ reference_article;
 		System::String^ designation;
 		System::Decimal^ prix_produit_HT;
-		System::Decimal^ taux_TVA;
+		System::Int32^ taux_TVA;
 		System::String^ couleur;
 		System::String^ produit;
 		System::Int32^ id_commande;
@@ -24,7 +24,7 @@ namespace NS_Comp_Mappage9
 		void setReference_article(System::String^);
 		void setDesignation(System::String^);
 		void setPrix_produitHT(System::Decimal^);
-		void setTaux_tva(System::Decimal^);
+		void setTaux_tva(System::Int32^);
 		void setCouleur(System::String^);
 		void setProduit(System::String^);
 		void setId_commande(System::Int32^);
